Adds validation rules module
The validationRules module will hold all the built-in validation rules
for the engine. These will be merged with any custom validation rules a
user might provide as well.

That module also holds the constructors for valid object and error
object that are added to the validator's "store"

utils/validator holds a recursive function that combs the form schema
for validation entries and builds the validation list on rendering of the form.