Renames validator.js from utils more appropriately & tests
The tests already existed for gatherValidations, I believe they were
removed for expediency in testing for a change in another branch and the
commented out tests got committed anyway.

Also because the validator is a component now instead of being wholly a
module, it didn't make sense for this file to remain being called
"validator.js", so renamend to match its actual contents more closely.