Adds validate all function
Validate all function iterates through all the form values that are
currently held by the form controller. Whenever it comes across an entry
that matches something that needs to be validated, it runs the single
Validation on it.

Currently the validate all fn does not catch required fields which have
not been touched and therefore have no valid OR invalid validation
entry.

There are two hurdles to overcome in this as I see it-
1) what entry will be made in the validationState, to which path will it
match? Or is there a generic *not found* required validation object
2) any field components will have to look for this entry based on name
instead of the usual check through path