Changes top level schema def
Removed the idea of a schema at all, and just had the top level
FormController take a prop called "properties" and treated it as though
it were an object. The props are then passed down to an object rendering
component so everything may proceed as normal.

The "name" field should be optional, though encouraged. here it will not
error, but it will use the title (camel cased) to be the name field if
one isn't provided.