Moved deciding which node into the object node
Changed the main node layer to the object node.
There was a very uneccessary step in the FormNode component just routing
through to which node should handle the field based on the type. The
object node can handle that, because the object node only renders other
nodes, there isn't much else going on there.

Objects also feel like where the other nodes should be travelling
through anyway. It's the holder of all the ndoe data, and the properties
themselves.