Adds Array Type field and tests
Started to implement the Array field type.
Made the decision to go with a class component instead of a functional
component because there were a few variables and functions needed within
the closure. This started to feel too much like state, and it wasn't
hard to imagine the component requiring something like that in the
future.

Instead of trying to skirt the issue and shove some state into a
functional component, changed to class.