Moves value handling into HOC
Removes the valueManager into its own HOC. This keeps a stark separation
of concerns, the valueManager keeps track of the form data, and the
FormController acts as a holding point for all these different behaviors
(validation, values) and starts the rendering process.

An argument could be made that the FormController itself should be a
stateless functional component.