Adds submitShape functionality
submitShape is a function that will scan through a form's object schema
and return an array of what fields need to be added to the params for
submit. Each field is represented by an object that holds the field's
name and validations. Extra properties are added for arrays and objects:
their type and children.

Adds the subtype of objects "object:nosubmit". Will act like an object
in every other way, but in submitting, it will be ignored as if it's
children are flat to the same scope it exists in. This should be used
for objects that are purely for layout or organization but do not need
to be submitted as params in the form's final submit.

This feature is added to avoid having to pass large objects or have
special cases for "gathering" the fields in order to submit. Without
the form scanning itself for what will be eventually submitted (and
validated before submission) special cases would have to be used to
"gather" the data and bundle it into an object for http params/etc..

- adds an optional parameter to the formcontroller component to be able
to pass in a custom submit button. this will likely be made more robust
in the near future.