Adds Submit Value gathering for singles & arrays
The form state that holds the values the user has inputted is holding
those values based on path, not name. This makes working with the form
while it's in memory much easier because each field value matches to a
unique identifier.

That approach however, leads to problems with submitting a form. Each of
those paths ends with the name of the field, but contains a lot more
information than wherever the form is submitting to will need.

Rather than require all backends to use the path syntax, an extra step
is required on submit to convert those path fields into information the
recieving end will be expecting.

In previous incarnations of this library/process there was either a
separate schema for what the shape of the eventual submitted params
would be, or would require a lengthy, typed in function that names all
the fields and gets their data. This leads to problems if the schema is
out of sync with that function, and plenty of room for typographic
errors to cause insidious bugs.

These functions read the form schema and create the submitted output
accordingly, rather than needing a separate schema.

This is possible because the form schema more closely resembles the data
it exists to capture. Any objects that are for organization or layout of
the form are labeled as such i the schema, and are ignored by the submit
gathering function.