Adds Widget passthrough for booleans
Adds the ability to accept widgets for boolean field types. Also adds
the styled boolean widget.

Styled booleans allow for more customization as far as how a boolean
value should be displayed by the UI. In the future I hope there will be
more ability to customize by adding in the ability to use a custom
component to display true/false. For now, it is limited to the styling
that can be done on the button element and its div.