Adds multiselect functionality
MultiSelect Widget
  Field Type: Array
  Field UI is made up of several checkboxes denoted by the options key in
  the form schema. Options should be strings, and are toggled into and
  out of the array with the checkbox inputs.

This commit also closes a bug found with the fieldRegistry and
valueManager that was not correctly saving fields/values that were
nested as objects or arrays.

The register function was not being properly handed down.

Apologies for this commit being so big, this PR was spread over a few
days and I lost track of some work.