Adds nodeWrapper component
Node wrapper will be placed around any node that is being rendered,
regardless of type or whether it's a widget or not.

the node wrapper is responsible for (at this time) conditional rendering
logic, label and description components, and the error handler to wrap
the node component.

Components and jsx specific to the node or widget are passsed into the
NodeWrapper as children.

When creating a new widget, it is not necessary to add the NodeWrapper,
it will be done in the node component for the corresponding type.