Removes excess css
Had trouble packaging up the css for the datepicker library. Compiled it
with sass on the command line, and added to the public folder.

TODO: Add to documentation that if you're using the datepicker widget,
you'll need to include the css.

Possibly will either make an alternative datepicker (not likely) or
research a cleaner way for someone using twenty questions to add in the
css to their project. look into: are there other packages that are
dependant on datepicker and how do they handle that?